fix: make a mapProducts util function
